Use dotenv.net for development secrets
Previously we had to manually edit the `launchSettings.json` file with the
development credentials for the CRM etc. This was suseptible to being
accidently commited into git (as we can't `gitignore` the `launchSettings.json`
file completely).

Instead, this commit introduces `dotnet.env` for managing development secrets
in an ignored `.env.development` file.