Log when searching around postcode
We are seeing an odd error in Sentry whereby we can't project between two
coordinates:

https://sentry.io/organizations/dfe-bat/issues/2063757538/?project=5276954&referrer=slack

The only thing I can think of that's causing this is a postcode is getting
geocoded to a latitude/longitude that falls outside the coordinate system we
have defined in `GeometryExtension`. This _shouldn't_ be possible as it covers
the whole of the United Kingdom, but perhaps there's a bad entry in the
database somewhere against a postcode that results in this error.

Adding the logging should help diagnose the issue where it happens again.

The only other place that could the cause is a postcode against a teaching
event; if this is the case it's going to be harder to diagnose so I'll cross
that bridge if/when we come to it.