Use correct coordinate system in GeocodeClientAdapter
When we were creating the coordinates in response to the Google location result
we weren't specifying a coordinate system. This meant that the coordinates
would use the default system (or no system, I'm not sure which) and when we
came to compare them in the radius search with a coordinate in the Wgs84
coordinate system it would raise an exception.

Updates to use the `NtsGeometryServices` factory to create the coordinate in
the correct system.