Allow GET requests to be cached privately for 5m
At the moment we monkey-patch the Ruby API client library to cache GET request
responses for 5 minutes. It was done this way originally as ASP.NET Core was
preventing us from setting `Cache-Control` headers on authenticated requests.
This no longer seems to be the case and the core team appear to be actively
working towards a less opinionated caching implementation.

Add a custom attribute for short term (5m), private response caching that is
safe to use on most GET responses. Applies it to:

- All /types endpoints
- All /privacy_policies endpoints
- GET /teaching_events endpoints (search and upcoming)