Log request details on validation failure
Currently when the API receives a request the appropriate request model
validator is ran automatically before the controller action is called. The side
effect of this is that the `LoggableAttribute` action filter is never ran for
these requests, meaning we don't have meaningful logs around what was in the
request payload.

This commit breaks out the logic to translate an `ActionExecutingContext` and
`ControllerActionDescriptor` into formatted, sanitized log messages to a new
utility class `LoggableMessageComposer`.

We then hook into the `PostConfigure` callback and override the
`InvalidModelStateResponseFactory` to log out the request details before
calling the built in factory that formats and returns a helpful 400 response..