Optimize postcocde sanitization
We were using a `Regex` replace here which is a concise but inefficient way of
removing white space from a string. As this method is called once per row in
the CSV its ran around 1.7 million times per sync, so doing a
micro-optimization on the sanitization offers a performance boost.