Add ability to generate long-lived access tokens for candidates
We want to be able to email candidates with a URL that links them straight into
a pre-filled mailing list sign up form. In order to do this we need to be able
to generate long-lived access tokens that are stored against a `contact` record
in Dynamics. We can then send these out to candidates and use the token to
retrieve their details from the API before presenting the form to them.

Add endpoint for generating long-lived access tokens against a candidate.
Tokens are stored in Dynamics and are not returned, so the only way to expose
them is via an email from the CRM itself.