Exchange long lived access token for mailing list sign up
When a long lived access token is emailed out to a candidate we need a way for
the Rails app to exchange this for the candidate information, so that the sign
up form can be pre-filled.

Add method to exchange a long lived access token for mailing list sign up.