Sync archived teaching events
We are going to be adding a new page to the website that will allow users to
browse 'archived' online events - these are events that are now in the past, up
to 4 months old.

This commit updates the sync logic to query and cache events as far back as 4
months ago.