Add metrics for TOTP generation/verification
We want to be able to track how many TOTPs are generated and, out of those,
which are successfully verified (and, similarly, which are generated but never
verified).

The metrics collect the `candidate_id` and the token so that we can attribute
metrics to specific users without storing PII. The outcome of the verification
process is also stored as a label, so we can determine if the correct TOTP was
provided.