Increment TOTP generated/verified metrics
When a TOTP is generated or verified we want to increment the corresponding
metric. In order to link OTPs to a candidate the TOTP interfaces have been
expanded to include a `candidateId`.

We also store the TOTP itself in the metric label so that we can see when the
candidate verified the TOTP and _which_ one was verified (if they generated
multiple). We store these in plain text as they are within our internal
influxdb instance, only last 15 minutes and are useless without the candidate
information that we don't store.