Add TOTP metrics to engagement row
We are now sending metrics for generated/verified Timed One Time Passwords. This enables us to see how many we are generating vs how many are being verified, in addition to how many times a user is failing to enter a valid TOTP.

Add Grafana panels to visualize the new metrics.