Extract candidate consent into SubscriptionManager
When a candidate signs up for a service they get a particular level of consent.
To clean up the request models this logic has been extracted into the new
`SubscriptionManager` class.

In addition, after talking to the CRM team we are also going to opt-out of bulk
postal mail/postal mail for event and mailing list candidates. Also, if a
candidate has already consented to something we will no longer opt them out due
to their latest subscription choice.