Add Client model/validations
Introduce a `Client` model that will represent a consumer of the API. A `Client` consists of:

- Name
- Description
- Role
- ApiKeyPrefix
- ApiKey

The `ApiKeyPrefix` will be used to construct the name of the environment variable that
will contain the client's `ApiKey`, in the format:

```
<api_key_prefix>_API_KEY_<environment>
```

For example, `GET_INTO_TEACHING_API_KEY_PRODUCTION`.

The `Role` attribute will be used to scope access to controllers/actions for
this particular client. A `Role` could be applied to multiple clients.