Add ClientManager to load API clients from YAML
When a request comes in to the API we need to be able to match the API key
against the associated client, so that we can authenticate and authorize the
request.

Add `ClientManager` that is responsible for loading a static list of clients
from YAML and provides a method to get a client associated with a given API key
(which will be used in the authentication handler).

The `Env` class has also been updated to return `Test` as the environment name
instead of `null`, which enables us to test matching a client to its
environment variable containing the API key.

Initially an `Administrator` client has been added, which will have access to
every endpoint (when we add the authorization logic to apply roles) and can be
used for testing locally and debugging in production.

Tests ensure that all clients in the YAML file are valid and there are no
duplicates.