Encorporate SharedSecretHandler into ApiClientHandler
Initially I thought I could register multiple authroization handlers, but it
turns out Web API only respects the final handler added.

To support the legacy `SharedSecretHandler` I've encorporated the logic into
the new `ApiClientHandler`; this way, if a `SHARED_SECRET` is used in the
`Authorization` header _or_ if a new API client key is used the request will be
authenticated and authorized with the `Admin` role.