Refactor actions to exchange access tokens for candidate data
We currently only have one way of the API clients obtaining candidate data,
which is via an access token that the user generates:

```
POST /path/to/resource/{accessToken}
```

Going forward we are also going to support 'magic links' as a way of
authenticating users. To that end, the endpoints are being refactored so we
will ultimately be left with:

```
POST /path/to/resource/exchange_access_token/{accessToken}
POST /path/to/resource/exchange_magic_link_token/{magicLinkToken}
```

The current endpoint is left in for backwards-compatibility and will be removed
once the clients have been updated.