Add magic link token fields to candidate
Add a field to store a magic link token against a candidate and the datetime
that the token was created, so that we can enforce token expiry.

Only the token field itself is mapped to the CRM for now; the others will be
mapped later on.

So that we can deploy this to production the `EntityFieldAttribute` has been
updated with an `Transient` option. If set, the `BaseModel` will not map the
fields to/from CRM `Entity` objects. This means we can safely deploy without
the CRM having been updated to include the new field.