Update CrmService to support magic links
We need to be able to query multiple candidates by their id from the CRM, so
that we can bulk-process the magic link token generation (the CRM will be
sending emails to thousands of candidates).

In addition, we need to be able to take a magic link from a request and find
the corresponding candidate record. This method returns a collection of
candidates in the edge case that a magic link has been duplicated during
generation or when a contact record has been copied in the CRM. It has a
null/empty string guard statement to ensure potentially expensive queries are
avoided (most contacts _won't_ have a magic link token set).

The query method for `GetCandidate` had to be updated from LINQ to a
QueryExpression as the Dynamics interpretter for LINQ does not support
`Contains`.