Add magic link token generator service
Add a service to generate secure, random 128 bit magic link tokens. It also has
a method that will exchange a valid token for the associated `Candidate`
record.

In the unlikely event that two candidates have been assigned the same magic
link token, the service will return `null` (ultimately forcing a re-generation
of the token for that candidate).