Set expiry date on magic link tokens
Change the `createdAt` date to be `expiresAt` and set to 48 hours in the future
when creating a magic link token.

Check that the token is valid before responding with a pre-filled
`MailingListAddMember` model.

If the magic link could not be matched to a candidate (or if it matches to
multiple candidates) we return `BadRequest` with a status of `Invalid`.

If a token was matched to a candidate but the expiry date has passed we return
`BadRequest` with a status of `Expired`.