Serialize enums as strings
Currently when an enum is returned the integer value is used. This seems to
cause an issue with swagger-codegen spitting out invalid Ruby code, for
example:

```
0 = 0.freeze
```

By serializing enums as string values I'm hoping it fixes the code generator.

We don't use enums in either of the client apps yet, so this change isn't
breaking.