Ignore model attribute mappings by environment
We currently have a number of new fields that are available in the dev/test CRM
environment but not production. In order to be able to deploy the code all the
way to production before the attributes are available we need to ignore them
from the mapping process.

Adds ability to exclude certain attributes from the mapping process depending
on environment.