Relex validation on returning teacher sign ups
If a candidate is already in the CRM the apps will now send existing data back
with the sign up payload (which will mean we can ultimately write `nil` values
to the CRM, as currently we have no way to differentiate between 'not set' and
'unset' fields).

A side-effect of this is that if an existing candidate is in the system with a
degree type and then goes through the TTA as a returner a load of validation
errors can occur due to the unexpected state (we don't ask new returners about
their degree status). To avoid this we relax the validation around certain
fields that trigger when the candidate has a degree - excluding returners.

The validation is getting pretty complex now; it may be worth relaxing some of
this in general in the API to avoid future problems, relying on the app-side
validation more.