Convert empty strings to null on serialize/deserialize operations
The CRM appears to have a number of existing candidates whereby the postcode,
for example, is an empty string `""` instead of `null`. This means when we
retrieve an existing record and pre-fill the store in the app we can send back
an invalid postcode to the API of an empty string instead of `null`. Some sign
ups are failing because the postcode is not correctly being identified as
optional/not specified. By serializing an empty string to `null` the behaviour
will be correct and the postcode validation will not be applied.