Add TrimStringJsonConverter
We have seen a few cases where the CRM has returned a postcode with a stray
space at the start or end. If the candidate matches-back and the postcode is
optional we will skip that step in the front-end and re-post the invalid
postcode back to the API, at which point the submission fails validation.

This converter will ensure all string values are trimmed of white space when
they exit/enter the API.