Trim strings coming back from the CRM
Some string values in the CRM have leading/trailing white space; this can be an
issue if they get ran through a strict validator (such as the postcode
validator).

By trimming strings returned from the CRM we can avoid potential issues with
validation later on.

If a string is empty we also want to treat it as null.