Add WebApplicationFactory to run app in test env
The WebApplicationFactory allows us to boot an instance of the application
running in the test environment and re-configure services that need to run
differently for tests.

The only change we make currently is to run the rate limit counters against an
in-memory store instead of Redis so that they are isolated between test runs.