Run DB migration earlier in startup
In some environments we check the database state before queueing jobs;
previously we were migrating the database _after_ this check, which throws an
exception if the database is empty (in need of the first migration being ran).

Moving the call to migrate the database earlier ensures it is ready in time.