Return NoResult instead of Failure when Authorization header missing
Some actions in the API don't require authentication (such as the
health-check). Currently, if a request doesn't contain an `Authorization`
header we return `AuthenticateResult.Failure` which marks the request as
unauthorized _and_ logs out an error.

Instead, we should be returning `Authenticate.NoResult` if the request can't be
authenticated due to missing `Authorization` header (returning
`AuthenticateResult.Failure` when the header is present but the API key is
invalid).

Removes manual logging as ASP.NET Core will log out the failure anyway, so we
were seeing duplicate log lines.