Stop inferring candidate type in TeacherTrainingAdviserSignUp
This is now passed explicitly from the client and marked as required in the
TeacherTrainingAdviserSignUp, so we no longer need to infer it.

Also updates `Candidate` to use the `TypeId` instead of inferring returning
teachers from if they have any `PastTeachingPositions`.

Various tests updated so that they make more sense (instead of setting
`SubjectTaughtId` or `PastTeachingPositions` to indicate a returner the
`TypeId` is now explicitly set).