Validate models mapped from CRM entities
When an entity comes back from the CRM we want to validate it so that we can
nullify any invalid attributes that the CRM returns.

An example would be if a `contact` entity has an invalid `address1_postalcode`;
if we mapped the invalid postcode to a `Candidate.AddressPostcode` and that was
then used in a match-back scenario the client application would end up posting
the invalid postcode back to the API and the submission would fail validation.

By sanitising the responses from the CRM we can ensure the API won't
inadvertently return invalid data to clients (and later receive it as part of a
submission).

We only sanitize the field attributes (not the relationships) - this is because
the related models get mapped recursively and their attributes will be
validated as part of that mapping process.