Track property changes on models
Implementing dirty tracking of model attributes in C# is hard. There are a
couple of ways you can do this; initially I tried using reflection to
dynamically override setters and track their usage. Whilst this is possible, it
requires making all properties `virtual` and injecting a proxy class that
handles hijacking the setter methods. This had the side-effect of making a
bunch of tests fail (as FluentValidation validators are tied to the model
types) and also felt fickle; if a property was added and not virtual it would
be ignored.

Another approach is to use the `INotifyPropertyChanged` interface which is
typically used in C# MVVM patterns to drive view updates. This requires
manually overriding each accessor and calling the change notifier; again, not
ideal and prone to errors in the future if someone forgets to do this for their
new accessor.

This commit uses `INotifyPropertyChanged` in conjunction with Fody to allow the
setters to be modified at compile-time using the `Property` weaver. It means as
long as our class implements `INotifyPropertyChanged` it will have the relevent
accessor code injected in to notify on property changes.

The `BaseModel` has been updated to include a `HashSet` of
`ChangedPropertyNames`, which we can use later on in order to ensure only
changed attributes are written to the CRM.

Hangfire automatically serializes payload objects using `Newtonsoft.Json`,
which doesn't deserialize the `ChangedPropertyNames` accurately (it modifies
the value as its deserializing, so you end up with all attributes being marked
as changed). To fix this the deserialization lifecycle is hooked into and
change tracking is disabled during that process. `System.Text.Json` appears to
deserialize the payloads fine, but I've included tests to ensure we know if
that changes in the future.