Only write changed attributes back to the CRM
Currently, the API will write all model attributes back to the CRM, skipping
`null` values entirely. This approach was taken due to having multiple request
models mapping back to the same `Candidate` model that gets persisted to the
CRM -- each request model contains a subset of `Candidate` attributes and we
don't want to `null`-out the attributes that weren't present in the request
model.

We had no way of distinguishing between a `Candidate` attribute that has been
set to `null` from a request model and a `Candidate` attribute that has been
unchanged by the request model.

To address this the base models have been updated to include change tracking
information. Now when we map from a request model to, say, a `Candidate`, we
maintain a `HashSet` of the changed property names. The mapping code now
references these and skips any attributes that have not been changed (also
writing any that have been changed and are now `null`).

When the `Candidate` is mapped to a CRM `Entity` the entity will maintain its
own change tracking and ensures only changed properties are written back to the
CRM.