Don't change channel for existing candidates
When an existing candidate has matched-back and gone through one of our 3
transactions we do not want to change their `ChannelId` in the CRM. Previously
we would set this to `null` and `null` values were not written to the CRM, but
now we write `null` values we need to instead not change the `ChannelId` (we
only write changed attributes to the CRM now).

Includes minor change to ensure a candidate qualification is not created on
event sign up unless they opt-in to the mailing list and specify their degree
subject.

Removes redundant setting of education phase for returner when setting past
teaching positions; this is already set in another method of this class.