Fix bug in deserializing ChangedPropertyNames
Initially I thought System.Text.Json was correctly deserializing
`ChangedPropertyNames` by default, but it turns out when you get more complex
objects with nesting and inheritance it can break due to the way the serializer
works. It was also dependent on the order of the attributes when serialized.

In order to consistently serialize/deserialize models with change tracking this
commit:

- Turns on `IgnoreNullValues` - if the serializer encounters a `null` we don't
  want it to deserialize the attribute and modify `ChangedPropertyNames`. If a
`null` _was_ changed it will already be present in the list when deserialized
anyway. This will also keep the Hangifre payload length down and make it easier
to interpret.

- Turn change tracking off during deserialization. We serialize
  `ChangedPropertyNames` so we don't want the deserializer to inadvertently
effect the values.

- Improves test coverage around deserialization to ensure various
  orders/combinations work correctly.