Manually serialize/deserialize UpsertCandidateJob payload
If we pass an object to a Hangfire job the library serializes/deserilizes it
for us using Newtonsoft.Json. This library is buggy when it comes to
deserializing the change tracking information and we also use System.Text.Json
everywhere else in the application, so it makes sense to do it manually/using
the same library. It also means we can have test coverage for the
serialization/deserialization process.