Remove custom model logging
We introduced the `[Loggable]` `[LogRequests]` and `[SensitiveData]` attributes
so that we could log out request/response details and model details in the
event of a validation error.

These attributes use `Newtonsoft.Json` to serialize/deserialize and they are no
longer compatible with our models (due to an extension that lets use use
`OnDeserializing` and `OnDeserialized` with `System.Text.Json`). The logging
also hasn't been working accurately for failed requests.

We have decent logging in the Rails apps and there is a ticket open in Trello
to improve the logging in the API in general, so I'm removing this code for now
and we can look to re-introduce it, hopefully as a better solution, in the
future.