Remove high cardinality metric labels
It turns out that using high cardinality labels against metrics is bad as each
combination is recorded as its own time-series, which is expensive (lots of
different label values == lots of different time series).

We do this in a couple of metrics and for the most part the labels aren't
actually used when it comes to graphing or alerting on the metrics, so this
commit removes them.