Use Newtonsoft.Json for Hangfire payloads
In an ideal world we would use `System.Text.Json` for all serialization in the
API. Whilst we got to this state recently it has a few drawbacks:

1. Dahomey.Json extensions are required to pause change tracking during
se/deserialization. This brings various oddities into `System.Text.Json` and it
doesn't implement `System.Runtime.Serialization.OnDeserializing/OnDeserialized`
according to the spec (it misses the `StreamingContext` argument, which can
have knock on effects). It also brings in `AutoMapper` which is slow and in
general its difficult to follow code, making it iffy to rely on.

2. `System.Text.Json` doesn't support conditional serialization, meaning we
can't omit the `ChangedPropertyNames` attribute from API responses and include
them when serializing to Hangfire -- we have to always include them, which
isn't great as it pollutes the API output.

To avoid these drawbacks I've reverted the Hangfire payloads to be
se/deserialized using `Newtonsoft.Json`, which is a robust library in itself
and still well supported. It implements `OnDeserializing`/`OnDeserialized`
correctly and the only different appears to be that properties with default
values get included in the `ChangedPropertyNames` on deserializing, which is
fine and probably a better outcome anyway (we always want to consider these
properties 'changed').

This commit also utilises the `JsonExtensions` in a few other areas so we're
not manually constructing the same serialization options repeatedly.