Collect metric on in person event result counts
When a teaching event search is performed we want to know how many in-person
events are returned (as well as total events, including online).

Add a second metric that observes how many in-person events are returned in a
search result.

Add convenience method to `TeachingEvent` to determine if it is `InPerson`.