Log out request/response payloads
Add middleware that intercepts request/response payloads, redacts sensitive
data and them logs the request information.

So as not to upset other middleware in the chain the request/response body
streams are copied and rewound. `RecyclableMemoryStreamManager` is used for
performance/to avoid memory leaks.

Query parameters are assumed to be safe/not sensitive.