Add utilty class to redact sensitive data from JSON
We want to be able to log out the request/response bodies and Hangfire job
paylaods to aid in debugging when something goes wrong. While doing this we
need to ensure no personally identifiable information is sent to our logstash.

Add a `Redactor` utility class that redacts sensitive data (identified by
payload keys) from JSON.

Update README to explain why we have to use Newtonsoft.Json for this.