Add AppSettings model
We want to be able to flag the CRM as unavailable when it has been
intentionally taken offline by the CRM team (for an update, for example).

Storing this in Postgres felt like overkill, so there is now a simple
`AppSettings` model that is backed by Redis. It has a single writable attribute
to set how long we want to pause the CRM integration for (this is a `DateTime`
so that we can auto-resume the integration and avoid potential user-error).

The idea is that the CRM will call an endpoint in the API and flag itself as
being unavailable before an update is carried out. We can then prevent any
calls to the CRM, which would otherwise result in errors.