Allow CRM integration to be paused/resumed
Adds two new operations endpoints for pausing/resuming the integration with the
CRM.

This could have been more rest-ful by accepting a payload with the integration
date, but I imagine if we get more `AppSettings` that can be modified
externally we will want tight access control, which is currently at the routing
layer so I think having individual endpoints is the way to go for now.

Updates the health check operation to return `integration-paused` as the CRM
status if applicable (the overall state will also be moved to `degraded`).