Don't call CRM if the integration is paused
If we know the CRM is unavailable we should not call out to it (as it results
in an error we can do nothing about).

Updates to prevent access tokens being generated when the CRM is offline
(instead we return `404` not found which results in the user being treated as a
new candidate/mirroring the current behavior sans error).

It is still possible for the API to call out to the CRM as part of the token verification process if:

- The user generates a token
- The CRM goes offline
- The user attempts to verify the token

This should be unlikely to happen often and, whilst we could potentially handle
it gracefully, its going to be a fair amount of work. I think we should see
if/how often it occurs before committing to handling it explicitly.