Skip sync/magic link generation jobs when CRM integration paused
When the CRM is unavailable we are unable to perform a sync/generate magic link
tokens.

Ideally we would prevent the jobs from being picked up at all when the API is
in this state, but Hangfire doesn't support this out of the box and the
workarounds aren't ideal. It also means we don't have to monitor the
`AppSettings` state to know when to resume processing jobs.

If the CRM is offline these jobs succeed immediately and log out that they have
been effectively 'skipped'.