Abort UpsertCandidateJob when CRM integration paused
Ideally we would fail the job silently or prevent it from being picked up by
pausing the queue, but Hangfire does not support either of these.

Instead, we raise an error with a clear explanation so when we are alerted we
don't have to investigate the source of the error.