Fix flakey rate limit specs
The rate limit tests were re-using the `WebApplicationFactory` across all tests
by way of the `ClassFixture`. As our database is setup/torn down per-test we
need to recreate the factory as well, otherwise the factory is pointing to a
now destroyed database and we get an error (depending on if the tests run
faster than the database is cleaned up or not, sometimes it would work other
times it would fail).