Add contract test
Adds support for running contract tests output from the TTA service.

The tests run against a test instance of the application, which mocks out the
CRM interactions and allows us to query objects that would have been sent to
the CRM - they are serialized and saved as the output contracts of the API (for
the CRM to ingest into their contract tests).

The application state is loaded prior to running tests, so that the data is
consistent with what was available to the TTA service (and the current time is
forzen consistently as well).

The contract output is committed and the snapshot is compared on subsequent
runs. If the calls to the CRM change then the snapshot will differ and the test
will fail (we would then commit the updated snapshot assuming the change is
expected).